Fix the reported amount of GPU memory physically available for GPPU with video mem larger than 4Gb

+++ libraries/shared/src/GPUIdent.h
@@ -14,17 +14,19 @@
 #ifndef hifi_GPUIdent_h
 #define hifi_GPUIdent_h
 
+#include <cstdint>
+
 class GPUIdent
 {
 public:
-    unsigned int getMemory() { return _dedicatedMemoryMB; }
+    uint64_t getMemory() { return _dedicatedMemoryMB; }
     QString getName() { return _name; }
     QString getDriver() { return _driver; }
     bool isValid() { return _isValid; }
     // E.g., GPUIdent::getInstance()->getMemory();
     static GPUIdent* getInstance(const QString& vendor = "", const QString& renderer = "") { return _instance.ensureQuery(vendor, renderer); }
 private:
-    uint _dedicatedMemoryMB { 0 };
+    uint64_t _dedicatedMemoryMB { 0 };
     QString _name { "" };
     QString _driver { "" };
     bool _isQueried { false };
